wip: working state
This commit brings the encoder value reading to a working state, however
it uses a combination of GUI setup (generated code) and custom setup
logic to properly configure and initialise the GPIO pins and Timer
peripheral to work correctly with a rotary encoder.

The setup logic should be moved exclusively into user code so nothing
important becomes buried.